+++ network/api/serializers.py
@@ -2,47 +2,7 @@ from django.conf import settings
 from django.contrib.sites.models import Site
 from rest_framework import serializers
 
-from network.base.models import (Antenna, Data, Observation, Satellite,
-                                 Station, Transponder)
-
-
-class AntennaSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Antenna
-        fields = ('frequency', 'band', 'antenna_type')
-
-
-class StationSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Station
-        fields = ('owner', 'name', 'image', 'alt', 'lat', 'lng',
-                  'antenna', 'featured_date', 'id')
-
-    image = serializers.SerializerMethodField('image_url')
-
-    def image_url(self, obj):
-        site = Site.objects.get_current()
-        return '{}{}{}'.format(site, settings.MEDIA_URL, obj.image)
-
-
-class SatelliteSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Satellite
-        fields = ('norad_cat_id', 'name')
-
-
-class TransponderSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Transponder
-        fields = ('description', 'alive', 'uplink_low', 'uplink_high',
-                  'downlink_low', 'downlink_high', 'mode', 'invert',
-                  'baud', 'satellite')
-
-
-class ObservationSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Observation
-        fields = ('satellite', 'transponder', 'author', 'start', 'end')
+from network.base.models import Data
 
 
 class DataSerializer(serializers.ModelSerializer):
@@ -64,7 +24,7 @@ class JobSerializer(serializers.ModelSerializer):
                   'frequency')
 
     def get_frequency(self, obj):
-        return obj.observation.transponder.downlink_low
+        return obj.observation.transmitter.downlink_low
 
     def get_tle0(self, obj):
         return obj.observation.satellite.tle0