+++ network/base/management/commands/fetch_data.py
@@ -4,7 +4,7 @@ import urllib2
 from django.core.management.base import BaseCommand, CommandError
 from django.conf import settings
 
-from network.base.models import Satellite, Transponder
+from network.base.models import Satellite, Transmitter
 
 
 class Command(BaseCommand):
@@ -13,11 +13,11 @@ class Command(BaseCommand):
     def handle(self, *args, **options):
         apiurl = settings.DB_API_ENDPOINT
         satellites_url = "{0}satellites".format(apiurl)
-        transponders_url = "{0}transponders".format(apiurl)
+        transmitters_url = "{0}transmitters".format(apiurl)
         self.stdout.write("Fetching from: {0}".format(satellites_url))
         try:
             satellites = urllib2.urlopen(satellites_url).read()
-            transponders = urllib2.urlopen(transponders_url).read()
+            transmitters = urllib2.urlopen(transmitters_url).read()
         except:
             raise CommandError('API is unreachable')
 
@@ -32,23 +32,23 @@ class Command(BaseCommand):
                 sat.save()
                 self.stdout.write('Satellite {0}-{1} added'.format(norad_cat_id, name))
 
-        for transponder in json.loads(transponders):
-            norad_cat_id = transponder['norad_cat_id']
-            uuid = transponder['uuid']
-            description = transponder['description']
+        for transmitter in json.loads(transmitters):
+            norad_cat_id = transmitter['norad_cat_id']
+            uuid = transmitter['uuid']
+            description = transmitter['description']
 
             try:
                 sat = Satellite.objects.get(norad_cat_id=norad_cat_id)
             except:
                 self.stdout.write('Satellite {0} not present'.format(norad_cat_id))
-            transponder.pop('norad_cat_id')
+            transmitter.pop('norad_cat_id')
             try:
-                existing_transponder = Transponder.objects.get(uuid=uuid)
-                existing_transponder.__dict__.update(transponder)
-                existing_transponder.satellite = sat
-                self.stdout.write('Transponder {0}-{1} updated'.format(uuid, description))
-            except Transponder.DoesNotExist:
-                new_transponder = Transponder.objects.create(**transponder)
-                new_transponder.satellite = sat
-                new_transponder.save()
-                self.stdout.write('Transponder {0}-{1} created'.format(uuid, description))
+                existing_transmitter = Transmitter.objects.get(uuid=uuid)
+                existing_transmitter.__dict__.update(transmitter)
+                existing_transmitter.satellite = sat
+                self.stdout.write('Transmitter {0}-{1} updated'.format(uuid, description))
+            except Transmitter.DoesNotExist:
+                new_transmitter = Transmitter.objects.create(**transmitter)
+                new_transmitter.satellite = sat
+                new_transmitter.save()
+                self.stdout.write('Transmitter {0}-{1} created'.format(uuid, description))